#78 [refactor] culture를 stadium 어그리거트에 통합하고 DDD구조로 변경한다.
