+++ build.gradle
@@ -73,8 +73,33 @@ dependencies {
 
 }
 
-def querydslDir = "src/main/generated"
+java {
+    toolchain {
+        languageVersion = JavaLanguageVersion.of(17) // Java 버전
+    }
+}
 
+tasks.named('compileJava', JavaCompile).configure {
+    options.annotationProcessorGeneratedSourcesDirectory = file("$build/generated/sources/annotationProcessor/java/main")
+}
+
+sourceSets {
+    main {
+        java {
+            srcDirs += "$build/generated/sources/annotationProcessor/java/main"
+        }
+    }
+}
+
+tasks.named('compileJava') {
+    doFirst {
+        // QClass 파일 자동 삭제
+        def querydslGeneratedDir = layout.buildDirectory.dir("generated/sources/annotationProcessor/java/main").get().asFile
+        if (querydslGeneratedDir.exists()) {
+            querydslGeneratedDir.deleteDir()
+        }
+    }
+}
 ext {
     snippetsDir = file('build/generated-snippets')
 }
