--- src/main/java/kusitms/backend/culture/domain/entity/Entertainment.java
@@ -1,44 +1,0 @@
-package kusitms.backend.culture.domain.entity;
-
-import jakarta.persistence.*;
-import kusitms.backend.culture.domain.enums.Boundary;
-import kusitms.backend.global.domain.BaseTimeEntity;
-import kusitms.backend.result.infra.converter.StringListConverter;
-import kusitms.backend.stadium.domain.model.Stadium;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Entertainment extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "entertainment_id", nullable = false)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "stadium_id", nullable = false)
-    private Stadium stadium;
-
-    @Column(nullable = false)
-    private String imgUrl;
-
-    @Enumerated(EnumType.STRING)
-    private Boundary boundary;
-
-    @Column(nullable = false)
-    private String name;
-
-    @Lob
-    @Convert(converter = StringListConverter.class)
-    private List<String> explanations;
-
-    @Lob
-    @Convert(converter = StringListConverter.class)
-    private List<String> tips;
-}
