--- src/main/java/kusitms/backend/culture/domain/repository/FoodRepository.java
@@ -1,8 +1,0 @@
-package kusitms.backend.culture.domain.repository;
-
-import kusitms.backend.culture.domain.entity.Food;
-import kusitms.backend.culture.domain.repository.custom.FoodRepositoryCustom;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface FoodRepository extends JpaRepository<Food, Long>, FoodRepositoryCustom {
-}
