--- src/main/java/kusitms/backend/culture/status/FoodErrorStatus.java
@@ -1,37 +1,0 @@
-package kusitms.backend.culture.status;
-
-import kusitms.backend.global.code.BaseErrorCode;
-import kusitms.backend.global.dto.ErrorReasonDto;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-
-@Getter
-@RequiredArgsConstructor
-public enum FoodErrorStatus implements BaseErrorCode {
-    _BAD_REQUEST_BOUNDARY(HttpStatus.BAD_REQUEST, "FOOD-001", "잘못된 영역입니다. 내부 또는 외부로 입력해주세요."),
-    _BAD_REQUEST_COURSE(HttpStatus.BAD_REQUEST, "FOOD-002", "잘못된 값입니다. 식사 또는 후식으로 입력해주세요.");
-
-    private final HttpStatus httpStatus;
-    private final String code;
-    private final String message;
-
-    @Override
-    public ErrorReasonDto getReason() {
-        return ErrorReasonDto.builder()
-                .isSuccess(false)
-                .code(code)
-                .message(message)
-                .build();
-    }
-
-    @Override
-    public ErrorReasonDto getReasonHttpStatus() {
-        return ErrorReasonDto.builder()
-                .isSuccess(false)
-                .httpStatus(httpStatus)
-                .code(code)
-                .message(message)
-                .build();
-    }
-}
