+++ src/main/java/kusitms/backend/stadium/domain/enums/Course.java
@@ -1,0 +1,24 @@
+package kusitms.backend.stadium.domain.enums;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+import java.util.Arrays;
+import java.util.Optional;
+
+@Getter
+@RequiredArgsConstructor
+public enum Course {
+
+    MEAL("식사"),
+    DESSERT("후식"),
+    TOTAL("전체");
+
+    private final String name;
+
+    public static Optional<Course> findByName(String name) {
+        return Arrays.stream(Course.values())
+                .filter(course -> course.getName().equals(name))
+                .findFirst();
+    }
+}
